// src/app/core/api/generated/uploads/uploads.service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import { HttpClient, HttpHeaders } from "@angular/common/http";
import type {
  HttpContext,
  HttpEvent,
  HttpParams,
  HttpResponse as AngularHttpResponse,
} from "@angular/common/http";

import { Injectable, inject } from "@angular/core";

import { Observable } from "rxjs";

import type {
  CreateUploadUrlDto,
  CreateUploadUrlResponseDto,
} from "../schemas";

interface HttpClientOptions {
  readonly headers?: HttpHeaders | Record<string, string | string[]>;
  readonly context?: HttpContext;
  readonly params?:
    | HttpParams
    | Record<
        string,
        string | number | boolean | Array<string | number | boolean>
      >;
  readonly reportProgress?: boolean;
  readonly withCredentials?: boolean;
  readonly credentials?: RequestCredentials;
  readonly keepalive?: boolean;
  readonly priority?: RequestPriority;
  readonly cache?: RequestCache;
  readonly mode?: RequestMode;
  readonly redirect?: RequestRedirect;
  readonly referrer?: string;
  readonly integrity?: string;
  readonly referrerPolicy?: ReferrerPolicy;
  readonly transferCache?: { includeHeaders?: string[] } | boolean;
}

@Injectable({ providedIn: "root" })
export class UploadsService {
  private readonly http = inject(HttpClient);
  uploadsControllerCreatePresignedUpload<TData = CreateUploadUrlResponseDto>(
    createUploadUrlDto: CreateUploadUrlDto,
    options?: HttpClientOptions & { observe?: "body" },
  ): Observable<TData>;
  uploadsControllerCreatePresignedUpload<TData = CreateUploadUrlResponseDto>(
    createUploadUrlDto: CreateUploadUrlDto,
    options?: HttpClientOptions & { observe: "events" },
  ): Observable<HttpEvent<TData>>;
  uploadsControllerCreatePresignedUpload<TData = CreateUploadUrlResponseDto>(
    createUploadUrlDto: CreateUploadUrlDto,
    options?: HttpClientOptions & { observe: "response" },
  ): Observable<AngularHttpResponse<TData>>;
  uploadsControllerCreatePresignedUpload<TData = CreateUploadUrlResponseDto>(
    createUploadUrlDto: CreateUploadUrlDto,
    options?: HttpClientOptions & { observe?: "body" | "events" | "response" },
  ): Observable<TData | HttpEvent<TData> | AngularHttpResponse<TData>> {
    if (options?.observe === "events") {
      return this.http.post<TData>(`/uploads/presign`, createUploadUrlDto, {
        ...(options as Omit<NonNullable<typeof options>, "observe">),
        observe: "events",
      });
    }

    if (options?.observe === "response") {
      return this.http.post<TData>(`/uploads/presign`, createUploadUrlDto, {
        ...(options as Omit<NonNullable<typeof options>, "observe">),
        observe: "response",
      });
    }

    return this.http.post<TData>(`/uploads/presign`, createUploadUrlDto, {
      ...(options as Omit<NonNullable<typeof options>, "observe">),
      observe: "body",
    });
  }
}