// src/app/core/api/generated/schemas/createUploadUrlResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { CreateUploadUrlResponseDtoHeaders } from "./createUploadUrlResponseDtoHeaders";

export interface CreateUploadUrlResponseDto {
  bucket: string;
  key: string;
  endpoint: string;
  contentType: string;
  folder: string;
  uploadUrl: string;
  method: string;
  headers: CreateUploadUrlResponseDtoHeaders;
  expiresIn: number;
}